Attempting to make this mofo run with some validation data.
