+++ validation/MFRTRT/MFRTRT_g_functions.csv
@@ -1,0 +1,76 @@
+-15.09127 ,-2.413131
+-14.956754,-2.347334
+-14.822238,-2.277856
+-14.687722,-2.204306
+-14.553207,-2.126403
+-14.418691,-2.043936
+-14.284175,-1.956748
+-14.14966 ,-1.864729
+-14.015144,-1.767818
+-13.880628,-1.666003
+-13.746113,-1.559325
+-13.611597,-1.447891
+-13.477081,-1.331875
+-13.342565,-1.211536
+-13.20805 ,-1.087226
+-13.073534,-0.959403
+-12.939018,-0.828628
+-12.804503,-0.695568
+-12.669987,-0.560985
+-12.535471,-0.425707
+-12.400955,-0.290607
+-12.26644 ,-0.15656
+-12.131924,-0.024406
+-11.997408,0.105089
+-11.862893,0.231271
+-11.728377,0.353623
+-11.593861,0.471782
+-11.459346,0.585544
+-11.32483 ,0.694859
+-11.190314,0.799809
+-11.055798,0.900586
+-10.921283,0.997462
+-10.786767,1.090764
+-10.652251,1.180841
+-10.517736,1.268046
+-10.38322 ,1.352718
+-10.248704,1.435174
+-10.114188,1.5157
+-9.979673 ,1.594548
+-9.845157 ,1.671942
+-9.710641 ,1.748074
+-9.576126 ,1.82311
+-9.44161  ,1.897193
+-9.307094 ,1.970447
+-9.172579 ,2.042979
+-9.038063 ,2.114878
+-8.903547 ,2.186224
+-8.769031 ,2.257086
+-8.634516 ,2.327523
+-8.5      ,2.585872
+-7.8      ,2.928872
+-7.2      ,3.224872
+-6.5      ,3.570872
+-5.9      ,3.867872
+-5.2      ,4.210872
+-4.5      ,4.549872
+-3.963    ,4.803872
+-3.27     ,5.126872
+-2.864    ,5.309872
+-2.577    ,5.435872
+-2.171    ,5.608872
+-1.884    ,5.725872
+-1.191    ,5.986872
+-0.497    ,6.208872
+-0.274    ,6.270872
+-0.051    ,6.326872
+0.196     ,6.381872
+0.419     ,6.424872
+0.642     ,6.462872
+0.873     ,6.495872
+1.112     ,6.523872
+1.335     ,6.544872
+1.679     ,6.569872
+2.028     ,6.587872
+2.275     ,6.598872
+3.003     ,6.619872
